Some random stuff
- add some network filter tabs
- add random selection buttons for trees & props
- show extra asset info in asset description pop-up
