+++ FindIt/Debugging.cs
@@ -4,7 +4,6 @@
 using System;
 using System.Text;
 using UnityEngine;
-using ColossalFramework;
 
 namespace FindIt
 {
