+++ FindIt/ModInfo.cs
@@ -14,7 +14,7 @@ namespace FindIt
 {
     public class ModInfo : IUserMod
     {
-        public const string version = "2.0.0";
+        public const string version = "2.0.1-beta1";
         public const bool isBeta = true;
 
         public string Name
